Splits between community and third party
