+++ src/pagePartials/badge/model/CreateCommunityBadgeModel.tsx
@@ -4,21 +4,21 @@ import { DEFAULT_COURT_ID } from '@/src/constants/common'
 import { contracts } from '@/src/contracts/contracts'
 import { useContractInstance } from '@/src/hooks/useContractInstance'
 import useTransaction from '@/src/hooks/useTransaction'
-import CreateWithSteps from '@/src/pagePartials/badge/model/CreateWithSteps'
-import { CreateModelSchemaType } from '@/src/pagePartials/badge/model/schema/CreateModelSchema'
+import CreateCommunityBadgeModelWithSteps from '@/src/pagePartials/badge/model/CreateCommunityBadgeModelWithSteps'
+import { CreateCommunityModelSchemaType } from '@/src/pagePartials/badge/model/schema/CreateCommunityModelSchema'
 import { BADGE_MODEL_TEXT_CONTRAST } from '@/src/pagePartials/badge/model/steps/uiBasics/BadgeModelUIBasics'
 import { useWeb3Connection } from '@/src/providers/web3ConnectionProvider'
 import { BadgeModelControllerName } from '@/types/badges/BadgeModel'
 import { TheBadgeModels__factory } from '@/types/generated/typechain'
 import { NextPageWithLayout } from '@/types/next'
 
-const CreateBadgeModel: NextPageWithLayout = () => {
+const CreateCommunityBadgeModel: NextPageWithLayout = () => {
   const { resetTxState, sendTx, state: transactionState } = useTransaction()
 
   const { address, appChainId, readOnlyAppProvider } = useWeb3Connection()
   const theBadgeModels = useContractInstance(TheBadgeModels__factory, 'TheBadgeModels')
 
-  const onSubmit = async (data: CreateModelSchemaType) => {
+  const onSubmit = async (data: CreateCommunityModelSchemaType) => {
     const {
       backgroundImage,
       badgeMetadataColumns,
@@ -102,8 +102,12 @@ const CreateBadgeModel: NextPageWithLayout = () => {
   }
 
   return (
-    <CreateWithSteps onSubmit={onSubmit} resetTxState={resetTxState} txState={transactionState} />
+    <CreateCommunityBadgeModelWithSteps
+      onSubmit={onSubmit}
+      resetTxState={resetTxState}
+      txState={transactionState}
+    />
   )
 }
 
-export default withPageGenericSuspense(CreateBadgeModel)
+export default withPageGenericSuspense(CreateCommunityBadgeModel)
