+++ src/pagePartials/badge/model/steps/community/strategy/PDFRequirementInput.tsx
